+++ demo_native/src/main.rs
@@ -7,7 +7,7 @@ fn main() -> Result<(), eframe::Error> {
     eframe::run_native(
         "MyApp",
         Default::default(),
-        Box::new(|cc| Box::new(MyApp::new(cc.egui_ctx.clone()))),
+        Box::new(|cc| Ok(Box::new(MyApp::new(cc.egui_ctx.clone())))),
     )
 }