+++ demo_web/src/main.rs
@@ -10,7 +10,7 @@ fn main() {
             .start(
                 "the_canvas_id", // hardcode it
                 web_options,
-                Box::new(|cc| Box::new(demo::MyApp::new(cc.egui_ctx.clone()))),
+                Box::new(|cc| Ok(Box::new(demo::MyApp::new(cc.egui_ctx.clone())))),
             )
             .await
             .expect("failed to start eframe");