+++ walkers/src/download.rs
@@ -33,7 +33,11 @@ impl Default for HttpOptions {
     fn default() -> Self {
         Self {
             cache: None,
-            user_agent: HeaderValue::from_static("Walkers"),
+            user_agent: HeaderValue::from_static(concat!(
+                env!("CARGO_PKG_NAME"),
+                "/",
+                env!("CARGO_PKG_VERSION"),
+            )),
         }
     }
 }